reference/requirements.txt
defusedxml==0.4.1
Django==1.8.9
django-bootstrap-toolkit==2.15.0
django-nose==1.4.3
django-tastypie==0.13.1
lxml==3.5.0
nose==1.3.7
python-dateutil==2.4.2
python-mimeparse==1.5.1
six==1.10.0
wheel==0.24.0
